+++ packages/core/event.d.ts
@@ -17,6 +17,7 @@ export default class EventWithInternals extends Event {
   constructor (errorClass: string, errorMessage: string, stacktrace: any[], handledState?: HandledState, originalError?: Error)
   _metadata: { [key: string]: any }
   _handledState: HandledState
+  _session?: Session
   toJSON(): {
     payloadVersion: '4'
     exceptions: Array<Error & { message: Error['errorMessage'] }>